Move scripts into SUPPORT directory and update tool paths
